// test/locators/login-page.enum.ts
export enum LoginPageEnum{
    IdentifierInput = 'identifier-input',
    PasswordInput = 'signin-password-input',
    ContinueButton = 'identifier-submit',
    LogInButton = 'signin-button',
    SignInWithGoogleButton = 'signin-with-google-button',
    SignInWithAppleButton = 'signin-with-apple-button',
    SignInWithCodeButton = 'signin-with-code-button',

    BusinessEmailInput = 'email-input',
    BusinessEmailSubmitButton = 'submit-email-button',

}